update chapter 01

+++ include/Color.cuh
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cmath>
+
 #include <Vec3.cuh>
 
 namespace TinyRT {
@@ -11,9 +13,9 @@ namespace TinyRT {
 
 		__host__ __device__ void gammaCorrect() {
 			// gamma-correct for gamma=2.0.
-			_elem[0] = sqrt(_elem[0]);
-			_elem[1] = sqrt(_elem[1]);
-			_elem[2] = sqrt(_elem[2]);
+			_elem[0] = sqrtf(_elem[0]);
+			_elem[1] = sqrtf(_elem[1]);
+			_elem[2] = sqrtf(_elem[2]);
 		}
 
 		__host__ __device__ float r() const { return _elem[0]; }
